requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.20.0
flask>=2.0.0
werkzeug>=2.0.0
tkinter
